move contract deployment hook to dedicated file
