+++ seth/client.go
@@ -1002,11 +1002,6 @@ func (cl *ContractLoader[T]) LoadContract(name string, address common.Address, a
 	return wrapperInitFn(address, cl.Client.Client)
 }
 
-type ContractDeploymentHooks struct {
-	Pre  func(auth *bind.TransactOpts, name string, abi abi.ABI, bytecode []byte, params ...interface{}) error
-	Post func(client *Client, tx *types.Transaction) error
-}
-
 func (m *Client) DeployContractWithHooks(hooks ContractDeploymentHooks, auth *bind.TransactOpts, name string, abi abi.ABI, bytecode []byte, params ...interface{}) (DeploymentData, error) {
 	L.Info().
 		Msgf("Started deploying %s contract", name)
@@ -1068,173 +1063,10 @@ func (m *Client) DeployContractWithHooks(hooks ContractDeploymentHooks, auth *bi
 // to the contract map, so that we can use that, when tracing transactions. It is suggested to use name identical to the name of the contract Solidity file.
 func (m *Client) DeployContract(auth *bind.TransactOpts, name string, abi abi.ABI, bytecode []byte, params ...interface{}) (DeploymentData, error) {
 	hooks := ContractDeploymentHooks{
-		Post: func(client *Client, tx *types.Transaction) error {
-			err := retry.Do(
-				func() error {
-					ctx, cancel := context.WithTimeout(context.Background(), m.Cfg.Network.TxnTimeout.Duration())
-					_, err := bind.WaitDeployed(ctx, m.Client, tx)
-					cancel()
-
-					// let's make sure that deployment transaction was successful, before retrying
-					if err != nil && !errors.Is(err, context.DeadlineExceeded) {
-						ctx, cancel := context.WithTimeout(context.Background(), m.Cfg.Network.TxnTimeout.Duration())
-						receipt, mineErr := bind.WaitMined(ctx, m.Client, tx)
-						if mineErr != nil {
-							cancel()
-							return mineErr
-						}
-						cancel()
-
-						if receipt.Status == 0 {
-							return errors.New("deployment transaction was reverted")
-						}
-					}
-
-					return err
-				}, retry.OnRetry(func(i uint, retryErr error) {
-					switch {
-					case errors.Is(retryErr, context.DeadlineExceeded):
-						replacementTx, replacementErr := prepareReplacementTransaction(m, tx)
-						if replacementErr != nil {
-							L.Debug().Str("Current error", retryErr.Error()).Str("Replacement error", replacementErr.Error()).Uint("Attempt", i+1).Msg("Failed to prepare replacement transaction for contract deployment. Retrying with the original one")
-							return
-						}
-						tx = replacementTx
-					default:
-						// do nothing, just wait again until it's mined
-					}
-					L.Debug().Str("Current error", retryErr.Error()).Uint("Attempt", i+1).Msg("Waiting for contract to be deployed")
-				}),
-				retry.DelayType(retry.FixedDelay),
-				// if gas bump retries are set to 0, we still want to retry 10 times, because what we will be retrying will be other errors (no code at address, etc.)
-				// downside is that if retries are enabled and their number is low other retry errors will be retried only that number of times
-				// (we could have custom logic for different retry count per error, but that seemed like an overkill, so it wasn't implemented)
-				retry.Attempts(func() uint {
-					if m.Cfg.GasBumpRetries() != 0 {
-						return m.Cfg.GasBumpRetries()
-					}
-					return 10
-				}()),
-				retry.RetryIf(func(err error) bool {
-					return strings.Contains(strings.ToLower(err.Error()), "no contract code at given address") ||
-						strings.Contains(strings.ToLower(err.Error()), "no contract code after deployment") ||
-						(m.Cfg.GasBumpRetries() != 0 && errors.Is(err, context.DeadlineExceeded))
-				}),
-			)
-
-			if err != nil {
-				// pass this specific error, so that Decode knows that it's not the actual revert reason
-				_, _ = m.Decode(tx, errors.New(ErrContractDeploymentFailed))
-
-				return wrapErrInMessageWithASuggestion(m.rewriteDeploymentError(err))
-			}
-
-			return nil
-		},
+		Post: RetryingPostDeployHook,
 	}
 
 	return m.DeployContractWithHooks(hooks, auth, name, abi, bytecode, params...)
-
-	// L.Info().
-	// 	Msgf("Started deploying %s contract", name)
-
-	// if auth.Context != nil {
-	// 	if err, ok := auth.Context.Value(ContextErrorKey{}).(error); ok {
-	// 		return DeploymentData{}, errors.Wrapf(err, "aborted contract deployment for %s, because context passed in transaction options had an error set", name)
-	// 	}
-	// }
-
-	// address, tx, contract, err := bind.DeployContract(auth, abi, bytecode, m.Client, params...)
-	// if err != nil {
-	// 	return DeploymentData{}, wrapErrInMessageWithASuggestion(err)
-	// }
-
-	// L.Info().
-	// 	Str("Address", address.Hex()).
-	// 	Str("TXHash", tx.Hash().Hex()).
-	// 	Msgf("Waiting for %s contract deployment to finish", name)
-
-	// m.ContractAddressToNameMap.AddContract(address.Hex(), name)
-
-	// if _, ok := m.ContractStore.GetABI(name); !ok {
-	// 	m.ContractStore.AddABI(name, abi)
-	// }
-
-	// // retry is needed both for gas bumping and for waiting for deployment to finish (sometimes there's no code at address the first time we check)
-	// if err := retry.Do(
-	// 	func() error {
-	// 		ctx, cancel := context.WithTimeout(context.Background(), m.Cfg.Network.TxnTimeout.Duration())
-	// 		_, err := bind.WaitDeployed(ctx, m.Client, tx)
-	// 		cancel()
-
-	// 		// let's make sure that deployment transaction was successful, before retrying
-	// 		if err != nil && !errors.Is(err, context.DeadlineExceeded) {
-	// 			ctx, cancel := context.WithTimeout(context.Background(), m.Cfg.Network.TxnTimeout.Duration())
-	// 			receipt, mineErr := bind.WaitMined(ctx, m.Client, tx)
-	// 			if mineErr != nil {
-	// 				cancel()
-	// 				return mineErr
-	// 			}
-	// 			cancel()
-
-	// 			if receipt.Status == 0 {
-	// 				return errors.New("deployment transaction was reverted")
-	// 			}
-	// 		}
-
-	// 		return err
-	// 	}, retry.OnRetry(func(i uint, retryErr error) {
-	// 		switch {
-	// 		case errors.Is(retryErr, context.DeadlineExceeded):
-	// 			replacementTx, replacementErr := prepareReplacementTransaction(m, tx)
-	// 			if replacementErr != nil {
-	// 				L.Debug().Str("Current error", retryErr.Error()).Str("Replacement error", replacementErr.Error()).Uint("Attempt", i+1).Msg("Failed to prepare replacement transaction for contract deployment. Retrying with the original one")
-	// 				return
-	// 			}
-	// 			tx = replacementTx
-	// 		default:
-	// 			// do nothing, just wait again until it's mined
-	// 		}
-	// 		L.Debug().Str("Current error", retryErr.Error()).Uint("Attempt", i+1).Msg("Waiting for contract to be deployed")
-	// 	}),
-	// 	retry.DelayType(retry.FixedDelay),
-	// 	// if gas bump retries are set to 0, we still want to retry 10 times, because what we will be retrying will be other errors (no code at address, etc.)
-	// 	// downside is that if retries are enabled and their number is low other retry errors will be retried only that number of times
-	// 	// (we could have custom logic for different retry count per error, but that seemed like an overkill, so it wasn't implemented)
-	// 	retry.Attempts(func() uint {
-	// 		if m.Cfg.GasBumpRetries() != 0 {
-	// 			return m.Cfg.GasBumpRetries()
-	// 		}
-	// 		return 10
-	// 	}()),
-	// 	retry.RetryIf(func(err error) bool {
-	// 		return strings.Contains(strings.ToLower(err.Error()), "no contract code at given address") ||
-	// 			strings.Contains(strings.ToLower(err.Error()), "no contract code after deployment") ||
-	// 			(m.Cfg.GasBumpRetries() != 0 && errors.Is(err, context.DeadlineExceeded))
-	// 	}),
-	// ); err != nil {
-	// 	// pass this specific error, so that Decode knows that it's not the actual revert reason
-	// 	_, _ = m.Decode(tx, errors.New(ErrContractDeploymentFailed))
-
-	// 	return DeploymentData{}, wrapErrInMessageWithASuggestion(m.rewriteDeploymentError(err))
-	// }
-
-	// L.Info().
-	// 	Str("Address", address.Hex()).
-	// 	Str("TXHash", tx.Hash().Hex()).
-	// 	Msgf("Deployed %s contract", name)
-
-	// if !m.Cfg.ShouldSaveDeployedContractMap() {
-	// 	return DeploymentData{Address: address, Transaction: tx, BoundContract: contract}, nil
-	// }
-
-	// if err := SaveDeployedContract(m.Cfg.ContractMapFile, name, address.Hex()); err != nil {
-	// 	L.Warn().
-	// 		Err(err).
-	// 		Msg("Failed to save deployed contract address to file")
-	// }
-
-	// return DeploymentData{Address: address, Transaction: tx, BoundContract: contract}, nil
 }
 
 // rewriteDeploymentError makes some known errors more human friendly
